refactor: move config package out of tracker/, remove ulimit functionality
