--- tracker/config/ulimit_darwin.go
@@ -1,8 +1,0 @@
-//go:build darwin
-// +build darwin
-
-package config
-
-func ulimitBugged() bool {
-	return true
-}
