+++ tracker/http/announce.go
@@ -6,8 +6,8 @@ import (
 	"net/netip"
 	"strconv"
 
+	"github.com/crimist/trakx/config"
 	"github.com/crimist/trakx/pools"
-	"github.com/crimist/trakx/tracker/config"
 	"github.com/crimist/trakx/tracker/stats"
 	"github.com/crimist/trakx/tracker/storage"
 )
