+++ tracker/http/announce_bench_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/cbeuw/connutil"
-	"github.com/crimist/trakx/tracker/config"
+	"github.com/crimist/trakx/config"
 	"github.com/crimist/trakx/tracker/storage"
 )
