+++ tracker/storage/map/memory.go
@@ -8,7 +8,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/crimist/trakx/tracker/config"
+	"github.com/crimist/trakx/config"
 	"github.com/crimist/trakx/tracker/storage"
 	"github.com/crimist/trakx/tracker/utils"
 	"github.com/pkg/errors"
