+++ tracker/udp/conndb.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/crimist/trakx/tracker/config"
+	"github.com/crimist/trakx/config"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
