+++ tracker/udp/error.go
@@ -1,7 +1,7 @@
 package udp
 
 import (
-	"github.com/crimist/trakx/tracker/config"
+	"github.com/crimist/trakx/config"
 	"github.com/crimist/trakx/tracker/stats"
 	"github.com/crimist/trakx/tracker/udp/protocol"
 	"go.uber.org/zap"
