+++ tracker/udp/udptracker.go
@@ -10,7 +10,7 @@ import (
 	"net/netip"
 	"sync"
 
-	"github.com/crimist/trakx/tracker/config"
+	"github.com/crimist/trakx/config"
 	"github.com/crimist/trakx/tracker/stats"
 	"github.com/crimist/trakx/tracker/storage"
 	"github.com/crimist/trakx/tracker/udp/protocol"
